Refactors in pipeline test runner (#1735)
Reuse JSON unmarshaler.
Extract test case execution to its own function.
Rename some files according to code conventions.
