+++ internal/testrunner/runners/pipeline/testresult.go
@@ -7,9 +7,7 @@ package pipeline
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 
@@ -107,11 +105,11 @@ func compareJsonNumbers(a, b json.Number) bool {
 
 func diffJson(want, got []byte, specVersion semver.Version) (string, error) {
 	var gotVal, wantVal interface{}
-	err := jsonUnmarshalUsingNumber(want, &wantVal)
+	err := formatter.JSONUnmarshalUsingNumber(want, &wantVal)
 	if err != nil {
 		return "", fmt.Errorf("invalid want value: %w", err)
 	}
-	err = jsonUnmarshalUsingNumber(got, &gotVal)
+	err = formatter.JSONUnmarshalUsingNumber(got, &gotVal)
 	if err != nil {
 		return "", fmt.Errorf("invalid got value: %w", err)
 	}
@@ -173,7 +171,7 @@ func adjustTestResult(result *testResult, config *testConfig) (*testResult, erro
 		}
 
 		var m common.MapStr
-		err := jsonUnmarshalUsingNumber(event, &m)
+		err := formatter.JSONUnmarshalUsingNumber(event, &m)
 		if err != nil {
 			return nil, fmt.Errorf("can't unmarshal event: %s: %w", string(event), err)
 		}
@@ -200,7 +198,7 @@ func adjustTestResult(result *testResult, config *testConfig) (*testResult, erro
 
 func unmarshalTestResult(body []byte) (*testResult, error) {
 	var trd testResultDefinition
-	err := jsonUnmarshalUsingNumber(body, &trd)
+	err := formatter.JSONUnmarshalUsingNumber(body, &trd)
 	if err != nil {
 		return nil, fmt.Errorf("unmarshalling test result failed: %w", err)
 	}
@@ -208,28 +206,6 @@ func unmarshalTestResult(body []byte) (*testResult, error) {
 	var tr testResult
 	tr.events = append(tr.events, trd.Expected...)
 	return &tr, nil
-}
-
-// jsonUnmarshalUsingNumber is a drop-in replacement for json.Unmarshal that
-// does not default to unmarshaling numeric values to float64 in order to
-// prevent low bit truncation of values greater than 1<<53.
-// See https://golang.org/cl/6202068 for details.
-func jsonUnmarshalUsingNumber(data []byte, v interface{}) error {
-	dec := json.NewDecoder(bytes.NewReader(data))
-	dec.UseNumber()
-	err := dec.Decode(v)
-	if err != nil {
-		if err == io.EOF {
-			return errors.New("unexpected end of JSON input")
-		}
-		return err
-	}
-	// Make sure there is no more data after the message
-	// to approximate json.Unmarshal's behaviour.
-	if dec.More() {
-		return fmt.Errorf("more data after top-level value")
-	}
-	return nil
 }
 
 func marshalTestResultDefinition(result *testResult, specVersion semver.Version) ([]byte, error) {
@@ -253,7 +229,7 @@ func marshalNormalizedJSON(v interface{}, specVersion semver.Version) ([]byte, e
 	}
 
 	var obj interface{}
-	err = jsonUnmarshalUsingNumber(msg, &obj)
+	err = formatter.JSONUnmarshalUsingNumber(msg, &obj)
 	if err != nil {
 		return msg, err
 	}
